Students Marks Sum
You are given an array of integers, marks, denoting the marks scored by students in a class.
The alternating elements marks0, marks2, marks4, and so on denote the marks of boys.
Similarly, marks1, marks3, marks5, and so on denote the marks of girls.
The array name marks, works as a pointer which stores the base address of that array. In other words, marks contains the address where marks0 is stored in the memory.
For example, let marks = [3,2,5] and marks stores 0x7fff9575c05f. Then, 0x7fff9575c05f is the memory address of marks0.