Small Triangles, Large Triangles
You are given n, triangles, specifically, their sides ai, bi and ci. Print them in the same style but sorted by their areas from the smallest one to the largest one. It is guaranteed that all the areas are different. The best way to calculate a volume of the triangle with sides a, b and c is Heron's formula:
s = √(P * (P - a) * (P - b) * (P - c)) where P = (a+b+c)/2.